+++ src/AppBanner/index.tsx
@@ -10,7 +10,7 @@ import type { TrackClick } from '../utils/tracking';
 
 import { AppStore } from '../assets/app-store';
 import { PlayStore } from '../assets/play-store';
-import { BannerStyleData } from '../styles/colorData';
+import { ColorStylesData } from '../styles/colorData';
 import { selfServeStyles } from '../styles/bannerCommon';
 import { useEscapeShortcut, OnCloseClick, CLOSE_BUTTON_ID } from '../bannerCommon/bannerActions';
 import { styles } from './styles';
@@ -18,7 +18,7 @@ import { styles } from './styles';
 import { canRender, COMPONENT_NAME } from './canRender';
 export { COMPONENT_NAME };
 
-const defaultColors: BannerStyleData = {
+const defaultColors: ColorStylesData = {
     styleBackground: '#ebe8e8',
     styleHeader: `#333333`,
     styleBody: '#666',
@@ -27,10 +27,6 @@ const defaultColors: BannerStyleData = {
     styleButton: '#ffffff',
     styleButtonBackground: '#052962',
     styleButtonHover: '#234b8a',
-    styleReminderButton: '#121212',
-    styleReminderButtonBackground: '#ededed',
-    styleReminderButtonHover: '#dcdcdc',
-    styleReminderAnimation: '#707070',
     styleClose: `#333333`,
     styleCloseBackground: '#ebe8e8',
     styleCloseHover: '#ffd213',
