+++ src/BannerWithLink/index.tsx
@@ -2,12 +2,12 @@ import React, { useState } from 'react';
 import { Button, LinkButton, SvgCross } from '@guardian/source-react-components';
 import { useEscapeShortcut, OnCloseClick, CLOSE_BUTTON_ID } from '../bannerCommon/bannerActions';
 import type { TrackClick } from '../utils/tracking';
-import { BannerStyleData } from '../styles/colorData';
+import { ColorStylesData } from '../styles/colorData';
 import { selfServeStyles } from '../styles/bannerCommon';
 import { canRender, COMPONENT_NAME } from './canRender';
 export { COMPONENT_NAME };
 
-const defaultColors: BannerStyleData = {
+const defaultColors: ColorStylesData = {
     styleBackground: '#ebe8e8',
     styleHeader: `#333333`,
     styleBody: '#666666',
@@ -16,10 +16,6 @@ const defaultColors: BannerStyleData = {
     styleButton: '#ffffff',
     styleButtonBackground: '#052962',
     styleButtonHover: '#234b8a',
-    styleReminderButton: '#121212',
-    styleReminderButtonBackground: '#ededed',
-    styleReminderButtonHover: '#dcdcdc',
-    styleReminderAnimation: '#707070',
     styleClose: `#333333`,
     styleCloseBackground: '#ebe8e8',
     styleCloseHover: '#e5e5e5',
