+++ src/styles/bannerCommon.ts
@@ -1,37 +1,12 @@
 import { css } from '@emotion/react';
 import type { Extras } from '../logic/types';
 import { neutral, space, from, until, body, headline } from '@guardian/source-foundations';
-import { BannerStyles, BannerStyleData, colorStringStyles } from './colorData';
+import { ColorStylesData, getColors } from './colorData';
 
 const imgHeight = '280';
 
-export const selfServeStyles = (userVals: Extras, defaults: BannerStyleData) => {
-    const style: BannerStyleData = Object.assign({}, defaults);
-    const defKeys: BannerStyles[] = Object.keys(defaults) as BannerStyles[];
-    const regex = new RegExp(/^#([A-Fa-f0-9]{6})$/);
-
-    defKeys.forEach((key) => {
-        const userVal = userVals[key];
-
-        // If user val is undefined, or an empty string, use default val
-        if (userVal == null || !userVal.length) {
-            return;
-        }
-
-        // Protect against CSS injection
-        const item = userVal.split(';')[0].trim();
-
-        // Protect against null or empty user strings
-        if (item == null || !item.length) {
-            return;
-        }
-
-        // Check for legitimate CSS color string values
-        // - we only support `#abcdef` color format
-        if (colorStringStyles.includes(key) && regex.test(item)) {
-            style[key] = item;
-        }
-    });
+export const selfServeStyles = (userVals: Extras, defaults: ColorStylesData) => {
+    const style = getColors(userVals, defaults);
 
     return {
         wrapper: css`
